+++ src/test/java/com/urswolfer/gerrit/client/rest/http/changes/parsers/ReviewResultParserTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2017 Urs Wolfer
+ * Copyright 2013-2021 Urs Wolfer
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.urswolfer.gerrit.client.rest.http.changes;
+package com.urswolfer.gerrit.client.rest.http.changes.parsers;
 
 import com.google.common.truth.Truth;
 import com.google.gerrit.extensions.api.changes.ReviewResult;
