+++ src/test/resources/com/urswolfer/gerrit/client/rest/http/changes/parsers/diff.json
